reformatting the code: I moved create_transaction function to class User
